Add tests for compliance
