--- src/test/java/com/czertainly/core/service/ComplianceServiceTest.java
@@ -1,7 +1,0 @@
-package com.czertainly.core.service;
-
-import com.czertainly.core.util.BaseSpringBootTest;
-
-class ComplianceServiceTest extends BaseSpringBootTest {
-
-}
